requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
pytesseract
numpy
rapidfuzz
